pyproject.toml
[tool.poetry]
name = "football-game"
version = "0.1.4"
description = "package python football game"
authors = ["Julio Martinez del Pino <juliomdp92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
sqlmodel = "0.0.*"
python-decouple = "3.8"
psycopg2-binary = "2.9.*"
pandas = "^2.2.0"

[tool.poetry.group.dev.dependencies]
coverage = { extras = ["toml"], version = "7.4.*" }
black = { version = "23.12.*", allow-prereleases = true }
pytest = "7.4.*"
pytest-html = "4.1.*"
pytest-cov = "4.1.*"
pylint = "3.0.*"
mkdocs = "1.5.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
